requirements.txt
gradio>=6.0.0
python-multipart
numpy
opencv-python
Pillow
trimesh>=4.0.0
scipy
pystray
pytz
networkx
lxml
svglib
reportlab
colormath
# Native Vector Engine dependencies
svgelements>=1.9.0
shapely>=2.0.0
mapbox-earcut
# FastAPI backend dependencies
fastapi
uvicorn
httpx
